requirements.txt
Django==2.2
Pillow==8.0.1
pytz==2020.1
sqlparse==0.4.1
